requirements.txt
flask-restplus
werkzeug==0.16.1
pymongo
requests
pylint
flask-cors
python-dotenv
dnspython